+++ app/src/components/Toast.tsx
@@ -1,43 +1,48 @@
 import * as React from 'react'
 import { createPortal } from 'react-dom'
 import { FiAlertCircle, FiAlertOctagon, FiAlertTriangle, FiX } from 'react-icons/fi'
-import { CommonContext } from '../contexts/CommonContext'
 import { capitalize } from '../globals'
 
-const Portal: React.FC = ({ children }) =>
+const Portal = ({ children }: { children: React.ReactNode }) =>
     createPortal(children, document.getElementById('toast') as HTMLDivElement)
 
 // * Toast's duration
 export const TOAST_DURATION = 2000
 
-const Toast = () => {
-    // contexts
-    const { toast, setToast } = React.useContext(CommonContext)
+interface ToastProps {
+    toast: any
+    setToast: React.Dispatch<any>
+}
 
-    // * Set toast's values from context' state
+const Toast = ({ toast, setToast }: ToastProps) => {
+    // * Clear Toast
+    React.useEffect(() => {
+        const timer = setTimeout(() => {
+            if (toast) setToast(null)
+        }, TOAST_DURATION)
+        return () => clearTimeout(timer)
+    })
 
     return (
         <Portal>
-            {toast && (
-                <aside className="toast">
-                    <div className="toast__header">
-                        <div className="toast__type">
-                            {toast.type === 'alert' && <FiAlertCircle />}
-                            {toast.type === 'error' && <FiAlertOctagon />}
-                            {toast.type === 'warning' && <FiAlertTriangle />}
-                            <span>{toast.type}</span>
-                        </div>
+            <aside className="toast">
+                <div className="toast__header">
+                    <div className="toast__type">
+                        {toast.type === 'alert' && <FiAlertCircle />}
+                        {toast.type === 'error' && <FiAlertOctagon />}
+                        {toast.type === 'warning' && <FiAlertTriangle />}
+                        <span>{toast.type}</span>
+                    </div>
 
-                        {/* close toast */}
-                        <button type="button" onClick={() => setToast(null)}>
-                            <FiX />
-                        </button>
-                    </div>
-                    <div className="toast__body">
-                        <p className="toast__message">{capitalize(toast.message)}</p>
-                    </div>
-                </aside>
-            )}
+                    {/* close toast */}
+                    <button type="button" onClick={() => setToast(null)}>
+                        <FiX />
+                    </button>
+                </div>
+                <div className="toast__body">
+                    <p className="toast__message">{capitalize(toast.message)}</p>
+                </div>
+            </aside>
         </Portal>
     )
 }