+++ changelog.md
@@ -4,3 +4,6 @@
 - Better messages on file download
 # 4.0.2
 - Nested Component definition refactor
+- Better messages on file download
+# 4.0.3
+- Decoupled Toast from Context